[이우섭] 0425 VMTextWidget 작업  - LottieWidget 제거, VMTextWidget 통합  - 1줄, 2줄 타이틀 각각 매칭

+++ lib/vm_sdk/types/text.dart
@@ -131,8 +131,8 @@ class TextData {
 
 class ExportedTextPNGSequenceData {
   String folderPath;
-  double width;
-  double height;
+  int width;
+  int height;
   double frameRate;
 
   ExportedTextPNGSequenceData(
